Separated open_and_save.c
