+++ src/ted.h
@@ -27,6 +27,11 @@ void setcolor(int c);
 // config_dialog.c
 void config_dialog();
 
+// open_and_save.c
+void savefile();
+void read_lines();
+void detect_linebreak();
+
 struct CFG
 {
     unsigned int tablen;
@@ -34,4 +39,17 @@ struct CFG
     unsigned char line_break_type : 2; // 0: LF  1: CRLF  2: CR
 };
 
+struct line
+{
+    unsigned int len;
+    unsigned char *data;
+    unsigned int length;
+    unsigned int real_length;
+    unsigned int ident;
+};
+
 extern struct CFG config;
+extern char *filename;
+extern unsigned int num_lines;
+extern struct line *lines;
+extern FILE *fp;
